Move bookbinder2 -> bookbinder
