+++ sotodlib/io/bookbinder.py
@@ -1,1260 +1,797 @@
-#!/usr/bin/env python3
-
-import os.path as op, os
-import so3g  # noqa: F401
+import so3g
 from spt3g import core
-import numpy as np
 import itertools
-import yaml
+import numpy as np
 from scipy.interpolate import interp1d
-from .load_smurf import split_ts_bits
-#
-class HWPFlags:
-    # STATIONARY is used to flag samples where data is being received by IRIG
-    # and the HWP agent, however the angle cannot be reconstructed because the HWP
-    # is not spinning.
-    STATIONARY = 2**0
-
-    # NO_DATA is used to flag samples where no HWP data is present in the
-    # hk files. This causes the `slow` field of the HWP g3 files to have
-    # `stable=False`.
-    NO_DATA = 2**1
+from scipy.signal import convolve
+from tqdm.auto import tqdm
+import os
+import yaml
+import logging
+import sys
+import shutil
+from sotodlib.site_pipeline.util import init_logger
 
 
-class _HKBundle():
+log = logging.getLogger('bookbinder')
+if not log.hasHandlers():
+    init_logger('bookbinder')
+
+
+def setup_logger(logfile=None):
     """
-    Buffer for Housekeeping data. Use add() to add data and rebundle() to output
-    data up to but not including flush_time.
+    This setups up a logger for bookbinder. If a logfile is passed, it will
+    write to that file as well as stdout. It is useful to create a one-off
+    logger here instead of using `getLogger` because it allows us to set
+    a separate log-file for each bookbinder instance.
     """
-    def __init__(self):
+    fmt = '%(asctime)s: %(message)s (%(levelname)s)'
+    log = logging.Logger('bookbinder', level=logging.DEBUG)
+
+    ch = logging.StreamHandler(sys.stdout)
+    formatter = logging.Formatter(fmt)
+    ch.setLevel(logging.INFO)
+    ch.setFormatter(formatter)
+    log.addHandler(ch)
+
+    if logfile is not None:
+        ch = logging.FileHandler(logfile)
+        ch.setLevel(logging.DEBUG)
+        ch.setFormatter(formatter)
+        log.addHandler(ch)
+
+    return log
+
+
+def get_frame_iter(files):
+    """
+    Returns a continuous iterator over frames for a list of files.
+    """
+    return itertools.chain(*[core.G3File(f) for f in files])
+
+
+def close_writer(writer):
+    """
+    Closes out a G3FileWriter with an end-processing frame. If None is passed,
+    this will not do anything.
+    """
+    if writer is None:
+        return
+    writer(core.G3Frame(core.G3FrameType.EndProcessing))
+
+
+def next_scan(it):
+    """
+    Returns the next Scan frame, along with any intermediate frames for an
+    iterator.
+    """
+    interm_frames = []
+    for frame in it:
+        if frame.type == core.G3FrameType.Scan:
+            return frame, interm_frames
+        interm_frames.append(frame)
+    return None, interm_frames
+
+
+class HKBlock:
+    def __init__(self, name):
+        self.name = name
+        self.data = {}
         self.times = []
-        self.data = None
+    
+    def process_frame(self, frame):
+        if self.name not in frame['block_names']:
+            return
+        idx = list(frame['block_names']).index(self.name)
+        block = frame['blocks'][idx]
 
-        self.azimuth_events = []
-        self.azimuth_starts = []
-        self.azimuth_stops = []
-        self.elevation_events = []
-        self.elevation_starts = []
-        self.elevation_stops = []
+        self.times.append(np.array(block.times) / core.G3Units.s)
+        for k, v in block.items():
+            if k not in self.data:
+                self.data[k] = []
+            self.data[k].append(v)
+    
+    def finalize(self):
+        if self.times:
+            self.times = np.hstack(self.times)
+            for k, v in self.data.items():
+                self.data[k] = np.hstack(v)
+        else:
+            self.times = np.array([], dtype=np.float64)
+            self.data = {}
 
-    def ready(self):
-        """
-        Returns True if a zero-crossing event exists
-        """
-        return len(self.azimuth_events) > 0
 
-    def add(self, b):
-        """
-        Cache the data from Housekeeping G3TimesampleMap b
-        """
-        self.times.extend(b.times)
-        if self.data is None:
-            self.data = {c: [] for c in b.keys()}
-        for c in b.keys():
-            self.data[c].extend(b[c])
-
-    def rebundle(self, flush_time):
-        """
-        Return the buffered Housekeeping G3TimesampleMap with all samples timestamped up to but
-        not including flush_time. Keep the samples from flush_time onward in buffer.
-
-        If there are no samples to bundle, None is returned.
-
-        Parameters
-        ----------
-        flush_time : G3Time
-            Buffered data up to (but not including) this time will be output, the rest kept
-
-        Returns
-        -------
-        output : G3SuperTimestream or None
-            Output signal timestream
-        """
-        if len(self.times) == 0:
-            return None
-
-        output = core.G3TimesampleMap()
-        output.times = core.G3VectorTime([t for t in self.times if t < flush_time])
-        self.times = [t for t in self.times if t >= flush_time]
-
-        for c in self.data.keys():
-            output[c] = core.G3VectorDouble(np.array(self.data[c][:len(output.times)]))
-            if len(output.times) < len(self.data[c]):
-                self.data[c] = self.data[c][len(output.times):]
-            else:
-                self.data[c] = np.array([])
-
-        return output
-
-class _SmurfBundle():
+class AncilProcessor:
     """
-    Buffer for SMuRF data. Use add() to add data and rebundle() to output data
-    up to but not including flush_time.
+    Processor for ancillary (ACU) data
+    
+    Params
+    --------
+    files : list
+        List of HK files to process
+
+    Attributes
+    -----------
+    blocks : dict
+        Dict containing ancillary blocks of co-sampled data, read from HK files.
+        This will be populated during preprocess.
+    times : np.ndarray
+        Timestamps for anc data. This will be populated after preprocess.
+    anc_frame_data : List[G3TimestreamMap]
+        List of G3TimestreamMaps saved for each bound frame. This will be
+        populated on bind and should be used to add copies of the anc data to
+        the detector frames. 
     """
-    def __init__(self):
-        self.times = []
-        self.signal = None
-        self.biases = None
-        self.primary = None
-        self.dtypes = {'data': np.int32, 'tes_biases': np.int32, 'primary': np.int64}
+    def __init__(self, files, book_id, log=None):
+        self.files = files
+        self.times = None
+        self.blocks = {
+            name: HKBlock(name) 
+            for name in ['ACU_broadcast', 'ACU_summary_output', 
+                         'HWPEncoder_freq']
+        }
+        self.anc_frame_data = None
+        self.out_files = []
+        self.book_id = book_id
+        self.preprocessed = False
 
-    def ready(self, flush_time):
+        if log is None:
+            self.log = logging.getLogger('bookbinder')
+        else:
+            self.log = log
+    
+    def preprocess(self):
         """
-        Returns True if the current frame has crossed the flush_time
+        Preprocesses HK data and populates the `data` and `times` objects.
         """
-        return len(self.times) > 0 and self.times[-1] >= flush_time
-
-    def add(self, f):
-        """
-        Cache the data from input SMuRF frame f
-
-        Parameters
-        ----------
-        f : G3Frame
-            Input SMuRF frame to be buffered
-        """
-        self.times.extend(f['data'].times)
-
-        if self.signal is None:
-            self.signal = {'names': f['data'].names, 'times': [], 'data': []}
-        assert (np.asarray(f['data'].names) == np.asarray(self.signal['names'])).all() # there's a bug where G3VectorString(x)==G3VectorString(x) returns False
-        assert f['data'].data.dtype == self.dtypes['data']
-        self.signal['times'].append(f['data'].times)
-        self.signal['data'].append(f['data'].data)
-
-        if 'tes_biases' in f.keys():
-            if self.biases is None:
-                self.biases = {'names': f['tes_biases'].names, 'times': [], 'data': []}
-            assert (np.asarray(f['tes_biases'].names) == np.asarray(self.biases['names'])).all()
-            assert f['tes_biases'].data.dtype == self.dtypes['tes_biases']
-            self.biases['times'].append(f['tes_biases'].times)
-            self.biases['data'].append(f['tes_biases'].data)
-
-
-        if 'primary' in f.keys():
-            if self.primary is None:
-                self.primary = {'names': f['primary'].names, 'times': [], 'data': []}
-            assert (np.asarray(f['primary'].names) == np.asarray(self.primary['names'])).all()
-            assert f['primary'].data.dtype == self.dtypes['primary']
-            self.primary['times'].append(f['primary'].times)
-            self.primary['data'].append(f['primary'].data)
-
-    def prepend(self, f):
-        """
-        Cache the data from input SMuRF frame f at the beginning of the buffer
-
-        Parameters
-        ----------
-        f : G3Frame
-            Input SMuRF frame to be buffered
-        """
-        if f['data'] is None:
+        if self.preprocessed:
             return
 
-        if len(self.times) == 0:
-            self.add(f)
-            return
-
-        # Original buffered data
-        t_orig = self.times
-        signal_orig = self.signal
-        biases_orig = self.biases
-        primary_orig = self.primary
-
-        # Clear buffers, add f, then append original data
-        self.__init__()
-        self.add(f)
-
-        self.times.extend(t_orig)
-
-        self.signal['times'] += signal_orig['times']
-        self.signal['data'] += signal_orig['data']
-        if biases_orig is not None:
-            self.biases['times'] += biases_orig['times']
-            self.biases['data'] += biases_orig['data']
-        if primary_orig is not None:
-            self.primary['times'] += primary_orig['times']
-            self.primary['data'] += primary_orig['data']
-
-    def rebundle(self, flush_time):
-        """
-        Return the buffered SMuRF G3SuperTimestreams with all samples timestamped up to but
-        not including flush_time. Keep the samples from flush_time onward in buffer.
-
-        If there are no samples to bundle, None is returned.
-
-        Parameters
-        ----------
-        flush_time : G3Time
-            Buffered data up to (but not including) this time will be output, the rest kept
-
-        Returns
-        -------
-        signalout : G3SuperTimestream
-            Output signal timestream
-        biasout : G3SuperTimestream or None
-            Output TES biases timestream
-        primout : G3SuperTimestream or None
-            Output primary timestream
-        """
-
-        # Timestamps to be output and kept in buffer, respectively
-        # (assumes signal, bias, and primary have same timestamps, which is a reasonable assumption)
-        tout       = [t for t in self.times if t < flush_time]
-        self.times = [t for t in self.times if t >= flush_time]
-
-        # Since G3SuperTimestreams cannot be shortened, the data must be copied to
-        # a new instance for rebundling of the buffer
-        signalstack = np.hstack(self.signal['data'])
-        signalout = so3g.G3SuperTimestream(self.signal['names'], tout, signalstack[:,:len(tout)])
-        self.signal = {'names': self.signal['names'], 'times': [self.times], 'data': [signalstack[:,len(tout):]]}
-
-        if self.biases is not None:
-            biasstack = np.hstack(self.biases['data'])
-            biasout = so3g.G3SuperTimestream(self.biases['names'], tout, biasstack[:,:len(tout)])
-            self.biases = {'names': self.biases['names'], 'times': [self.times], 'data': [biasstack[:,len(tout):]]}
-        else:
-            biasout = None
-
-        if self.primary is not None:
-            primstack = np.hstack(self.primary['data'])
-            primout = so3g.G3SuperTimestream(self.primary['names'], tout, primstack[:,:len(tout)])
-            self.primout = {'names': self.primary['names'], 'times': [self.times], 'data': [primstack[:,len(tout):]]}
-        else:
-            primout = None
-
-        return signalout, biasout, primout
-
-
-class FrameProcessor(object):
-    """
-    Module to process HK and SMuRF frames, adding to and rebundling from their buffers to
-    produce frames for the output Book.
-
-    Parameters
-    ----------
-    start_time : G3Time, optional
-        Start time for the output Book
-    end_time : G3Time, optional
-        End time for the output Book
-    smurf_timestamps : list, optional
-        Externally provided list of timestamps, used to check for missing samples in
-        current timestream
-    timing_system : bool
-        Whether the timing system was on during observation. If true, will attempt to use
-        timing counters to calculate correct timestamps; if false, will fall back to using
-        smurf-streamer recorded timestamps in the timestreams
-    flush_time : G3Time
-        Buffered data up to (but not including) this time will be output, the rest kept
-    maxlength : int
-        Maximum allowed length (in samples) of an output frame
-    """
-    def __init__(self, start_time=None, end_time=None, smurf_timestamps=None,
-                 timing_system=False, **config):
-        self.hkbundle = None
-        self.smbundle = None
-        self.hwp_loader = None
-        self.flush_time = None
-        self.maxlength = config.get("maxlength", 10000)
-        self.FLAGGED_SAMPLE_VALUE = config.get("flagged_sample_value", -1)
-        self.current_state = 0  # default to scan state
-        self.GAP_THRESHOLD = config.get("gap_threshold", 0.05)
-        self.BOOK_START_TIME = start_time
-        self.BOOK_END_TIME = end_time
-        self.timing_system = timing_system
-        self._frame_splits = []
-        self._hk_gaps = []
-        self._prev_smurf_sample = None
-        self._smurf_timestamps = smurf_timestamps
-        self._readout_ids = config.get("readout_ids", None)
-
-        # Ensure start and end times have the correct type
-        if self.BOOK_START_TIME is not None:
-            self.BOOK_START_TIME = core.G3Time(self.BOOK_START_TIME)
-        if self.BOOK_END_TIME is not None:
-            self.BOOK_END_TIME = core.G3Time(self.BOOK_END_TIME)
-
-    def ready(self):
-        """
-        Check if criterion passed in HK (for now, sign change in Az scan velocity data)
-        """
-        return self.hkbundle.ready() if (self.hkbundle is not None) else False
-
-    def locate_crossing_events(self, t, dy=0.001, min_gap=200):
-        """
-        Locate places where a noisy vector t changes sign: +ve, -ve, and 0. To handle noise,
-        "zero" is defined to be a threshold region from -dy to +dy. Thus t is said to have a
-        zero-crossing if it fully crosses the threshold region, with the actual crossing
-        approximated as halfway between entering and exiting the region. Entrances and exits
-        without a crossing are also identified (as "stops" and "starts" respectively).
-
-        Parameters
-        ----------
-        dy : float, optional
-            Amplitude of threshold region
-        min_gap : int, optional
-            Length of gap (in samples) longer than which events are considered separate
-
-        Returns
-        -------
-        events : list
-            Full list containing all zero-crossings, starts, and stops
-        starts : list
-            List of places where t exits the threshold region (subset of events)
-        stops : list
-            List of places where t enters the threshold region (subset of events)
-        """
-        tmin = np.min(t)
-        tmax = np.max(t)
-
-        if len(t) == 0:
-            return [], [], []
-
-        if np.sign(tmax) == np.sign(tmin):
-            return [], [], []
-
-        # If the data does not entirely cross the threshold region,
-        # do not consider it a sign change
-        if tmin > -dy and tmax < dy:
-            return [], [], []
-
-        # Find where the data crosses the lower and upper boundaries of the threshold region
-        c_lower = (np.where(np.sign(t[:-1]+dy) != np.sign(t[1:]+dy))[0] + 1)
-        c_upper = (np.where(np.sign(t[:-1]-dy) != np.sign(t[1:]-dy))[0] + 1)
-
-        # Classify each crossing as entering or exiting the threshold region
-        c_lower_enter = []; c_lower_exit = []
-        c_upper_enter = []; c_upper_exit = []
-
-        # Noise handling:
-        # If there are multiple crossings of the same boundary (upper or lower) in quick
-        # succession (i.e., less than min_gap), it is mostly likely due to noise. In this case,
-        # take the average of each group of crossings.
-        if len(c_lower) > 0:
-            spl = np.array_split(c_lower, np.where(np.ediff1d(c_lower) > min_gap)[0] + 1)
-            c_lower = np.array([int(np.ceil(np.mean(s))) for s in spl])
-            # Determine if this crossing is entering or exiting threshold region
-            for i, s in enumerate(spl):
-                if t[s[0]-1] < t[s[-1]+1]:
-                    c_lower_enter.append(c_lower[i])
-                else:
-                    c_lower_exit.append(c_lower[i])
-        if len(c_upper) > 0:
-            spu = np.array_split(c_upper, np.where(np.ediff1d(c_upper) > min_gap)[0] + 1)
-            c_upper = np.array([int(np.ceil(np.mean(s))) for s in spu])
-            # Determine if this crossing is entering or exiting threshold region
-            for i, s in enumerate(spu):
-                if t[s[0]-1] < t[s[-1]+1]:
-                    c_upper_exit.append(c_upper[i])
-                else:
-                    c_upper_enter.append(c_upper[i])
-
-        starts = np.sort(np.concatenate((c_lower_exit, c_upper_exit)))
-        stops = np.sort(np.concatenate((c_lower_enter, c_upper_enter)))
-        events = np.sort(np.concatenate((starts, stops)))
-
-        # Look for zero-crossings
-        zc = []
-        while len(c_lower) > 0 and len(c_upper) > 0:
-            # Crossing from -ve to +ve
-            if c_lower[0] < c_upper[0]:
-                b = c_lower[c_lower < c_upper[0]]
-                zc.append(int( np.ceil(np.mean([b[-1], c_upper[0]])) ))
-                c_lower = c_lower[len(b):]
-            # Crossing from +ve to -ve
-            elif c_upper[0] < c_lower[0]:
-                b = c_upper[c_upper < c_lower[0]]
-                zc.append(int( np.ceil(np.mean([b[-1], c_lower[0]])) ))
-                c_upper = c_upper[len(b):]
-
-        # Replace all upper and lower crossings that contain a zero-crossing in between with the
-        # zero-crossing itself, but ONLY if those three events happen in quick succession (i.e.,
-        # shorter than min_gap). Otherwise, they are separate events -- there is likely a stop
-        # state in between; in this case, do NOT perform the replacement.
-        for z in zc:
-            before_z = events[events < z]
-            after_z  = events[events > z]
-            if (after_z[0] - before_z[-1]) < min_gap:
-                events = np.concatenate((before_z[:-1], [z], after_z[1:]))
-
-        # If the last event is close to the end, a crossing of the upper or lower threshold
-        # boundaries may or may not be a significant event -- there is not enough remaining data
-        # to determine whether it is stopping or not. This will be clarified in the next iteration
-        # of the loop, when more HK data will have been added. (Or if there is no more HK data to
-        # follow, the loss in information by removing the last crossing is negligible.)
-        # On the other hand, if the last event is a zero-crossing, then that is unambiguous and
-        # should not be removed.
-        if (len(t) - events[-1] < min_gap) and events[-1] not in zc:
-            events = events[:-1]
-
-        # A similar problem occurs when the first boundary-crossing event is close to the
-        # beginning -- it could either be following a zero-crossing OR be entering/exiting a
-        # stopped state. In this case, we use previous data to disambiguate the two cases: if the
-        # telescope is currently in the scan state, a zero-crossing has just occurred so discard
-        # the event; otherwise, the threshold crossing indicates a change of state and should NOT
-        # be ignored.
-        if events[0] < min_gap and self.current_state == 0:
-            events = events[1:]
-
-        return events, starts, stops
-
-    def determine_state(self, v_az, v_el, dy=0.001):
-        """
-        Determine in which mode the telescope is operating:
-        - Mode 0: Scanning
-        - Mode 1: Stopped
-        - Mode 2: Slewing
-
-        Parameters
-        ----------
-        v_az : numpy.ndarray
-            Azimuth velocity
-        v_el : numpy.ndarray
-            Elevation velocity
-        dy : float, optional
-            Amplitude of the threshold region
-        """
-        if np.min(v_el) < -dy or np.max(v_el) > dy:
-            # If the el velocity exceeds the threshold region, the telescope is slewing
-            state = 2
-        elif np.min(v_az) < -dy or np.max(v_az) > dy:
-            # If the az velocity (but not el velocity) exceeds the threshold region, the telescope
-            # is scanning
-            state = 0
-        else:
-            # If the velocities lie entirely in the threshold region, the telescope is stopped
-            state = 1
-
-        self.current_state = state
-
-    def replace_times_and_trim_frame(self, f):
-        """
-        Replace the data timestamps with the most accurate available, then trim any samples
-        occurring before the Book start time and after the Book end time.
-
-        Parameters
-        ----------
-        f : G3Frame
-            Input frame to be processed
-
-        Returns
-        -------
-        trimmed_frame : G3Frame
-            Processed frame
-        """
-        if self.BOOK_START_TIME is not None and self.BOOK_END_TIME is not None:
-            assert self.BOOK_START_TIME <= self.BOOK_END_TIME
-
-        # Calculate the timestamps for the current frame, if available, or default to the
-        # recorded times in the 'data' field
-        t = get_timestamps(f, self.timing_system)
-
-        # Trim starting samples (if needed)
-        if self.BOOK_START_TIME is not None and t[0] < self.BOOK_START_TIME:
-            trimmed_frame_start = core.G3Frame(f.type)
-            t = core.G3VectorTime([_t for _t in t if _t >= self.BOOK_START_TIME])
-            for k in f.keys():
-                if k in ['data', 'tes_biases', 'primary']:
-                    trimmed_frame_start[k] = so3g.G3SuperTimestream(f[k].names, t, f[k].data[:,(len(f[k].times)-len(t)):])
-                else:
-                    trimmed_frame_start[k] = f[k]
-            trimmed_frame = trimmed_frame_start
-        else:
-            for k in f.keys():
-                if k in ['data', 'tes_biases', 'primary']:
-                    f[k].times = t
-            trimmed_frame = f
-
-        if len(trimmed_frame['data'].times) == 0:
-            return trimmed_frame
-
-        # Trim ending samples (if needed)
-        if self.BOOK_END_TIME is not None and t[-1] > self.BOOK_END_TIME:
-            trimmed_frame_end = core.G3Frame(trimmed_frame.type)
-            t = core.G3VectorTime([_t for _t in t if _t <= self.BOOK_END_TIME])
-            for k in f.keys():
-                if k in ['data', 'tes_biases', 'primary']:
-                    trimmed_frame_end[k] = so3g.G3SuperTimestream(f[k].names, t, f[k].data[:,:len(t)])
-                else:
-                    trimmed_frame_end[k] = trimmed_frame[k]
-            trimmed_frame = trimmed_frame_end
-
-        return trimmed_frame
-
-    def fill_in_missing_samples(self, data, flush_time, return_flags=False):
-        """
-        Locate and patch missing samples in data.
-
-        If a list of timestamps is provided, use it to fill in the missing samples (note: the
-        timestamps present in the input data must exactly match their counterparts in the list).
-        If not, determine the missing samples using an estimate of the sample interval.
-
-        Parameters
-        ----------
-        data : G3SuperTimestream
-            Input data to be screened for missing samples
-        flush_time : G3Time
-            Upper time limit (exclusive) for data to be returned
-        return_flags : bool, optional
-            Whether to return the list of flags indicating filled-in samples
-
-        Return
-        ------
-        dataout : G3SuperTimestream
-            Data G3SuperTimestream with filled-in samples wherever gaps are detected
-        flag_gap : G3VectorBool, optional
-            Array of boolean flags indicating filled-in samples that should not be used for
-            analysis. Set return_flags = True to return.
-        """
-        # There is no _prev_smurf_sample at the beginning of the book
-        if self._prev_smurf_sample is None:
-            prev_frame_end = self.BOOK_START_TIME.time - 1
-        else:
-            prev_frame_end = self._prev_smurf_sample.time
-        # Define a strict upper limit (exclusive) up until which to look/interpolate for missing samples
-        if self.BOOK_END_TIME is not None:
-            end_time = min(flush_time.time, self.BOOK_END_TIME.time + 1)
-        elif self._smurf_timestamps is not None:
-            end_time = min(flush_time.time, self._smurf_timestamps[-1] + 1)
-        else:
-            end_time = min(flush_time.time, data.times[-1].time + 1)
-
-        # Obtain list of filled-in timestamps, whether from provided list or estimate with fill_time_gaps
-        if self._smurf_timestamps is not None:
-            # trim the list of timestamps to the current frame
-            m_ = np.logical_and(self._smurf_timestamps < end_time, self._smurf_timestamps > prev_frame_end)
-            ts = self._smurf_timestamps[m_]
-        else:
-            # fallback when no timestamp is provided. We switch to estimate sample interval based on data
-            assert data.times[0].time >= prev_frame_end
-            # Use fill_time_gaps to fill in any missing samples in the middle, as well as
-            # at the beginning (between previous frame and current frame)
-            dt = np.median(np.diff(data.times))  # estimate sample interval
-            if self._prev_smurf_sample is None:
-                # Use estimated sample interval to fill backward to prev_frame_end = book start time - 1
-                # since book start time does not usually line up with sample interval
-                ts = np.append(np.flip(np.arange(data.times[0].time-dt, prev_frame_end, -dt, dtype=int)), data.times)
-                ts = fill_time_gaps(ts)
-            elif (data.times[0].time - prev_frame_end)/dt > 0.5:
-                ts = fill_time_gaps(np.insert(data.times, 0, prev_frame_end))[1:]
-            else:
-                ts = fill_time_gaps(np.asarray(data.times))
-            # Use estimated sample interval to fill until end_time. Not using fill_time_gaps because
-            # end_time is externally determined and does not usually line up with sample interval
-            ts = np.append(ts, np.arange(data.times[-1].time+dt, end_time, dt, dtype=int))
-
-        # Create new G3SuperTimestream with filled-in samples
-        if len(data.times) < len(ts):
-            dt = np.median(np.diff(ts))
-            i_missing, _ = find_missing_samples(ts, np.array(data.times), atol=dt/2)
-            m = np.ones(len(ts), dtype=bool)
-            m[i_missing] = False
-            assert np.sum(m) == len(data.times)
-            new_data = np.full((data.data.shape[0], len(ts)), self.FLAGGED_SAMPLE_VALUE, dtype=data.data.dtype)
-            new_data[:,m] = data.data
-            dataout = so3g.G3SuperTimestream(data.names, core.G3VectorTime(ts), new_data)
-            flag_gap = core.G3VectorBool(~m)[:self.maxlength]
-            print("{} missing samples detected".format(np.sum(~m)))
-        elif len(data.times) > len(ts):
-            raise ValueError("There are more samples in the timestream than in list provided")
-        else:
-            dataout = data
-            flag_gap = core.G3VectorBool(np.zeros(len(ts)))
-
-        if return_flags:
-            return dataout, flag_gap
-        else:
-            return dataout
-
-    def flush(self, flush_time=None):
-        """
-        Produce frames for the output Book, up to but not including flush_time
-
-        Parameters
-        ----------
-        flush_time : G3Time, optional
-            Buffered data up to (but not including) this time will be output, the rest kept
-
-        Returns
-        -------
-        output : list
-            List containing output frames
-        """
-        if flush_time == None:
-            flush_time = self.flush_time
-        output = []
-
-        smurf_data, smurf_bias, smurf_primary = self.smbundle.rebundle(flush_time)
-
-        # Fill in missing samples and write data to frame
-        f = core.G3Frame(core.G3FrameType.Scan)
-        smurf_data, flag_smurfgap = self.fill_in_missing_samples(smurf_data, flush_time, return_flags=True)
-        f['signal'], data_excess = split_supertimestream(smurf_data, self.maxlength)
-        fx = {'data': data_excess}  # note input data use 'data' instead of 'signal'
-        if smurf_bias is not None:
-            smurf_bias = self.fill_in_missing_samples(smurf_bias, flush_time)
-            f['tes_biases'], fx['tes_biases'] = split_supertimestream(smurf_bias, self.maxlength)
-        if smurf_primary is not None:
-            smurf_primary = self.fill_in_missing_samples(smurf_primary, flush_time)
-            f['primary'], fx['primary'] = split_supertimestream(smurf_primary, self.maxlength)
-        # Put any excess samples (those beyond max length) back in the buffer
-        self.smbundle.prepend(fx)
-        # Update (local) flush time if necessary
-        if data_excess is not None:
-            flush_time = data_excess.times[0]
-            self._frame_splits[-1] = flush_time  # update the most recently added split time
-        # Update last sample
-        self._prev_smurf_sample = f['signal'].times[-1]
-
-        # Gaps in SMuRF data
-        f['flag_smurfgap'] = core.G3VectorBool(flag_smurfgap[:len(f['signal'].times)])
-
-        # Ancillary data object to be written to frame
-        anc_data = core.G3TimesampleMap()
-        anc_data.times = f['signal'].times
-
-        ## Adds HWP data if present
-        if self.hwp_loader is not None:
-            ts = np.array(f['signal'].times) / core.G3Units.s
-            hwp_angle = self.hwp_loader.get_interpolated_data(ts)
-            anc_data['hwp_enc'] = core.G3VectorDouble(hwp_angle)
-
-        try:
-            hk_data = self.hkbundle.rebundle(flush_time)
-        except AttributeError:
-            # No az/el encoder data found in HK files
-            f['state'] = 3
-        else:
-            # Co-sampled (interpolated) az/el encoder data
-            cosamp_az = np.interp(f['signal'].times, hk_data.times, hk_data['Azimuth_Corrected'], left=np.nan, right=np.nan)
-            cosamp_el = np.interp(f['signal'].times, hk_data.times, hk_data['Elevation_Corrected'], left=np.nan, right=np.nan)
-
-            # Flag any samples falling within a gap in HK data
-            t = np.array([int(_t) for _t in f['signal'].times])
-            flag_hkgap = np.zeros(len(f['signal'].times))
-            for gap_start, gap_end in self._hk_gaps:
-                flag_hkgap = np.logical_or(flag_hkgap, np.logical_and(t > gap_start, t < gap_end))
-            cosamp_az = np.array([self.FLAGGED_SAMPLE_VALUE if flag_hkgap[i] else cosamp_az[i] for i in range(len(cosamp_az))])
-            cosamp_el = np.array([self.FLAGGED_SAMPLE_VALUE if flag_hkgap[i] else cosamp_el[i] for i in range(len(cosamp_el))])
-            f['flag_hkgap'] = core.G3VectorBool(flag_hkgap)
-
-            anc_data['az_enc'] = core.G3VectorDouble(cosamp_az)
-            anc_data['el_enc'] = core.G3VectorDouble(cosamp_el)
-            self.determine_state(hk_data['Azimuth_Velocity'], hk_data['Elevation_Velocity'])
-            f['state'] = self.current_state
-
-        finally:
-            # Write ancillary data to frame
-            f['ancil'] = anc_data
-            output += [f]
-
-        return output
-
-    def __call__(self, f):
-        """
-        Process a frame. Only Housekeeping and SMuRF frames will be manipulated;
-        others will be passed through untouched.
-
-        Parameters
-        ----------
-        f : G3Frame
-            Input frame to be processed
-
-        Returns
-        -------
-        output : list
-            List containing frames to be entered into output Book
-        """
-
-        if f.type != core.G3FrameType.Housekeeping and f.type != core.G3FrameType.Scan:
-            return [f]
-
-        if f.type == core.G3FrameType.Housekeeping:
-            if self.hkbundle is None:
-                self.hkbundle = _HKBundle()
-
-            try:
-                i = list(f['block_names']).index('ACU_position')
-            except ValueError:
-                print("'ACU_position' field not found in block names.")
-                return
-            self.hkbundle.add(f['blocks'][i])
-
-            self.hkbundle.data['Azimuth_Velocity'] = pos2vel(self.hkbundle.data['Azimuth_Corrected'])
-            self.hkbundle.data['Elevation_Velocity'] = pos2vel(self.hkbundle.data['Elevation_Corrected'])
-
-            azimuth_events = self.locate_crossing_events(self.hkbundle.data['Azimuth_Velocity'])
-            self.hkbundle.azimuth_events = [self.hkbundle.times[i] for i in azimuth_events[0]]
-            self.hkbundle.azimuth_starts = [self.hkbundle.times[i] for i in azimuth_events[1]]
-            self.hkbundle.azimuth_stops  = [self.hkbundle.times[i] for i in azimuth_events[2]]
-
-            elevation_events = self.locate_crossing_events(self.hkbundle.data['Elevation_Velocity'])
-            self.hkbundle.elevation_events = [self.hkbundle.times[i] for i in elevation_events[0]]
-            self.hkbundle.elevation_starts = [self.hkbundle.times[i] for i in elevation_events[1]]
-            self.hkbundle.elevation_stops  = [self.hkbundle.times[i] for i in elevation_events[2]]
-
-        if f.type == core.G3FrameType.Scan:
-            if self.smbundle is None:
-                self.smbundle = _SmurfBundle()
-
-            output = []
-
-            # Replace the times in 'data', 'tes_biases', and 'primary' with the correct timestamps
-            # and trim any samples occuring outside the specified start/end times
-            f = self.replace_times_and_trim_frame(f)
-            # Replace the channel names with readout IDs (if available)
-            f['data'].names = get_channel_names(f['data'], rids=self._readout_ids)
-
-            t = f['data'].times
-            # If all the samples have been trimmed, we can ignore this frame
-            if len(t) == 0:
-                return []
-
-            # Add current frame
-            self.smbundle.add(f)
-
-            # If the existing data exceeds the specified maximum length, split the frame into frames
-            # less than maxlength and record the frame split time(s)
-            while len(self.smbundle.times) > self.maxlength:
-                split_time = self.smbundle.times[self.maxlength]
-                # If flush_time occurs before split_time, exit loop for handling in following block
-                if self.flush_time is not None and self.flush_time <= split_time:
-                    break
-                self._frame_splits.append(split_time)
-                output += self.flush(split_time)
-            # If a frame split event has been reached, output a frame with samples up until flush_time
-            if self.flush_time is not None and self.smbundle.ready(self.flush_time):
-                output += self.flush()
-
-            return output
-
-class Bookbinder(object):
-    """
-    Bookbinder module. Co-samples and co-frames the provided Housekeeping (HK) and SMuRF data based
-    on telescope scan patterns to create the final, properly formatted archival Book.
-
-    If az/el encoder data exists in the HK file, Bookbinder identifies suitable frame split locations,
-    using this information. When no az/el encoder data is found in HK timestream, Bookbinder will run
-    in default mode, where frames are split once they reach the maximum length (Frameprocessor.MAXLENGTH).
-    HK data is co-sampled with SMuRF data before output.
-    """
-    def __init__(self, smurf_files, hk_files=None, out_root='.', book_id=None,
-                 session_id=None, stream_id=None, start_time=None, end_time=None,
-                 smurf_timestamps=None, timing_system=False, verbose=True,
-                 hwp_root=None, **config):
-        self._smurf_files = smurf_files
-        self._hk_files = hk_files
-        self._hwp_root = hwp_root
-        self._book_id = book_id
-        self._out_root = out_root
-        self._session_id = session_id
-        self._stream_id = stream_id
-        self._start_time = start_time
-        self._end_time = end_time
-        self._verbose = verbose
-        self._frame_splits_file = op.join(out_root, book_id, 'frame_splits.txt')
-        self._frame_splits = []
-        self._meta_file = op.join(out_root, book_id, f'M_index.yaml')
-        self.metadata = []
-        self.frame_num = 0
-        self.file_sample_counter = 0
-        self.prev_file_last_sample_num = 0
-        self.file_sample_ranges = [[0, 0]]
-        self.ofile_num = 0
-        self.default_mode = True  # True: time-based split; False: scan-based split
-        self.MAX_SAMPLES_TOTAL = int(config.get("max_samples_total", 1e9))
-        self.max_nchannels = int(config.get("max_nchannels", 1e3))
-        self.MAX_SAMPLES_PER_CHANNEL = self.MAX_SAMPLES_TOTAL // self.max_nchannels
-        self.DEFAULT_TIME = core.G3Time(1e18)  # 1e18 = 2286-11-20T17:46:40.000000000 (in the distant future)
-        self.OVERWRITE_ANCIL_FILE = config.get("overwrite_afile", False)
-        self.hwp_loader = None
-        self.ancil_writer = None
-        self.writer = None
-
-        if isinstance(self._hk_files, str):
-            self._hk_files = [self._hk_files]
-
-        # Verify start and end times
-        if self._start_time is not None:
-            self._start_time = core.G3Time(self._start_time)
-        if self._end_time is not None:
-            self._end_time = core.G3Time(self._end_time)
-
-        if self._start_time is not None and self._end_time is not None:
-            if int(self._start_time) >= int(self._end_time):
-                raise ValueError("Start time should be before end time." +
-                                 "\nStart time: " + str(self._start_time) +
-                                 "\nEnd time:   " + str(self._end_time))
-
-        # Set up the FrameProcessor
-        self.frameproc = FrameProcessor(start_time=self._start_time, end_time=self._end_time,
-                                        smurf_timestamps=smurf_timestamps,
-                                        timing_system=timing_system,
-                                        **config.get("frameproc_config", {}))
-
-        # Set up file I/O
-        if isinstance(self._smurf_files, list):
-            ifile = self._smurf_files.pop(0)
-            if self._verbose: print(f"Bookbinding {ifile}")
-            self.smurf_iter = core.G3File(ifile)
-
-            self.create_file_writers()
-
-    def close_file_writers(self):
-        """
-        Closes G3 file writers if they are open.
-        """
-        if self.ancil_writer is not None:
-            self.ancil_writer(core.G3Frame(core.G3FrameType.EndProcessing))
-        if self.writer is not None:
-            self.writer(core.G3Frame(core.G3FrameType.EndProcessing))
-
-    def create_file_writers(self):
-        """
-        Create the G3Writer instances for the output D-file and, if it doesn't already
-        exist, the A-file. Optionally, overwrite the existing A-file setting the
-        OVERWRITE_ANCIL_FILE attribute to True.
-        """
-
-        # Close existing file writers if open
-        self.close_file_writers()
-
-        out_bdir = op.join(self._out_root, self._book_id)
-        if not op.exists(out_bdir): os.makedirs(out_bdir)
-        ofile = op.join(out_bdir, f'D_{self._stream_id}_{self.ofile_num:03d}.g3')
-        if self._verbose: print(f"Writing {ofile}")
-        self.writer = core.G3Writer(ofile)
-
-        afile = op.join(out_bdir, f'A_ancil_{self.ofile_num:03d}.g3')
-        if not op.exists(afile) or self.OVERWRITE_ANCIL_FILE:
-            self.ancil_writer = core.G3Writer(afile)
-
-    def add_misc_data(self, f):
-        """
-        Insert miscellaneous metadata into frame
-
-        Parameters
-        ----------
-        f : G3Frame
-            Frame to be processed
-
-        Returns
-        -------
-        oframe : G3Frame
-            Output frame with metadata inserted
-        """
-        oframe = core.G3Frame(f.type)
-        oframe['book_id'] = self._book_id
-        for k in f.keys():
-            if k not in ['frame_num', 'session_id', 'stream_id', 'sostream_id', 'time']:
-                oframe[k] = f[k]
-        oframe['frame_num'] = core.G3Int(self.frame_num)
-        if self._session_id is not None:
-            oframe['session_id'] = core.G3Int(self._session_id)
-        if self._stream_id is not None:
-            oframe['stream_id'] = core.G3String(self._stream_id)
-        if oframe.type == core.G3FrameType.Scan:
-            range_start = self.prev_file_last_sample_num + self.file_sample_counter - len(f['signal'].times)
-            range_stop  = self.prev_file_last_sample_num + self.file_sample_counter
-            oframe['sample_range'] = core.G3VectorInt([range_start, range_stop])
-        return oframe
-
-    def write_frames(self, frames_list):
-        """
-        Write frames to file
-
-        Parameters
-        ----------
-        frames_list : list
-            List of frames to be written out
-        """
-        if not isinstance(frames_list, list):
-            frames_list = list(frames_list)
-
-        if len(frames_list) == 0: return
-
-        for f in frames_list:
-            # If the number of samples (per channel) exceeds the max allowed, create a new output file
-            if f.type == core.G3FrameType.Scan:
-                nsamples = len(f['signal'].times)   # number of samples in current frame
-                # if cumulative sample number exceeds max allowed, create new file to
-                # store the current frame and restart the sample number counter
-                if self.file_sample_counter + nsamples > self.MAX_SAMPLES_PER_CHANNEL:
-                    self.ofile_num += 1
-                    curr_file_last_sample_num = self.prev_file_last_sample_num + self.file_sample_counter
-                    self.prev_file_last_sample_num = curr_file_last_sample_num
-                    self.file_sample_counter = nsamples      # reset sample number to length of current frame
-                    self.file_sample_ranges.append([
-                        curr_file_last_sample_num,
-                        curr_file_last_sample_num + self.file_sample_counter
-                    ])
-                    self.frame_num = 0              # reset frame number to 0
-                    self.create_file_writers()
-                else:
-                    self.file_sample_counter += nsamples
-                    # keep file ranges up to date as we go
-                    self.file_sample_ranges[-1][1] = self.file_sample_counter + self.file_sample_ranges[-1][0]
-
-            # add misc metadata to output frame and write it out
-            oframe = self.add_misc_data(f)
-            self.writer.Process(oframe)
-            # if ancil_writer exists, write out the ancil frame
-            if self.ancil_writer is not None:
-                aframe = core.G3Frame(oframe.type)
-                if 'ancil' in oframe.keys():
-                    aframe['ancil'] = f['ancil']
-                if 'sample_range' in oframe.keys():
-                    aframe['sample_range'] = oframe['sample_range']
-                self.ancil_writer(aframe)
-            # update cumulative frame number
-            self.frame_num += 1
-
-    def find_frame_splits(self):
-        """
-        Determine the frame boundaries of the output Book
-
-        Returns
-        -------
-        frame_splits : list
-            List of timestamps where the frames should be split in the output Book
-        """
-        frame_splits = []
-
-        if self.default_mode:
-            frame_splits += [self.DEFAULT_TIME]  # no split, leave frameprocessor to decide based on maxlength
-            return frame_splits
-
-        # Assign a value to each event based on what occurs at that time: 0 for a zero-crossing,
-        # +/-1 for an azimuth start/stop, and +/-2 for an elevation start/stop. The cumulative
-        # sum of the resulting vector marks the state of the telescope at each event.
-        event_times = np.unique(np.concatenate((self.frameproc.hkbundle.azimuth_events, self.frameproc.hkbundle.elevation_events)))
-        values = []
-        for e in event_times:
-            # Note that multiple events can occur at the same event time, e.g., both az and el
-            # can start moving at the same time
-            v = 0
-            if e in self.frameproc.hkbundle.azimuth_starts:
-                v += 1
-            if e in self.frameproc.hkbundle.azimuth_stops:
-                v -= 1
-            if e in self.frameproc.hkbundle.elevation_starts:
-                v += 2
-            if e in self.frameproc.hkbundle.elevation_stops:
-                v -= 2
-            values.append(v)
-        seq = np.cumsum(values)
-
-        # If the housekeeping file starts while the telescope is moving, a stop event will occur
-        # prior to a start event, resulting in negative values in the seq vector.
-        # In that case, shift every value up by the lowest value.
-        starts_while_moving = False
-        m = np.min(seq)
-        if m < 0:
-            starts_while_moving = True
-            seq -= m
-
-        # Each time the seq returns to zero, the telescope is coming to a complete stop (in both
-        # az and el). Split the seq vector at these places.
-        sp = np.where(seq == 0)[0] + 1
-        sps = np.array_split(seq, sp)
-        spi = np.array_split(range(len(seq)), sp)
-
-        # Determine the times where the frame should be split
-        for i, s in enumerate(sps):
-            if len(s) == 0:
+        self.log.info("Preprocessing HK Data")
+        frame_iter = get_frame_iter(self.files)
+
+        for fr in frame_iter:
+            if fr['hkagg_type'] != 2:
                 continue
-            if np.max(s) >= 2:
-                # The telescope is slewing; only record the start (if it exists) and
-                # the stop of the slew
-                if i == 0 and starts_while_moving:
-                    slew_stop = event_times[spi[i][-1]]
-                    frame_splits += [slew_stop]
-                else:
-                    slew_start = event_times[spi[i][0]]
-                    slew_stop  = event_times[spi[i][-1]]
-                    frame_splits += [slew_start, slew_stop]
+
+            for block in self.blocks.values():
+                block.process_frame(fr)
+
+        for block in self.blocks.values():
+            block.finalize()
+        self.preprocessed = True
+
+    
+    def bind(self, outdir, times, frame_idxs, file_idxs):
+        """
+        Binds ancillary data.
+
+        Params
+        ----------
+        outdir : path
+            Path where files should be written
+        times : np.ndarray
+            Array of timestamps to write to book
+        frame_idxs : np.ndarray
+            Array mapping sample (in times) to output frame index
+        file_idxs : np.ndarray
+            Array mapping output frame idx to output file idx
+        """
+        self.log.info("Binding ancillary data")
+
+        # Handle file writers
+        writer = None
+        cur_file_idx = None
+        out_files = []
+
+        az, el, boresight, corotation = None, None, None, None
+        block = self.blocks['ACU_broadcast']
+        if 'Corrected_Azimuth' in block.data:
+            az = np.interp(times, block.times, block.data['Corrected_Azimuth'])
+            el = np.interp(times, block.times, block.data['Corrected_Elevation'])
+        if 'Corrected_Corotation' in block.data:
+            boresight = np.interp(times, block.times, block.data['Corrected_Boresight'])
+        if 'Corrected_Corotation' in block.data:
+            corotation = np.interp(times, block.times, block.data['Corrected_Corotation'])
+
+        anc_frame_data = []
+        for oframe_idx in np.unique(frame_idxs):
+            # Update file writer if starting a new output file
+            if file_idxs[oframe_idx] != cur_file_idx:
+                close_writer(writer)
+                cur_file_idx = file_idxs[oframe_idx]
+                fname = os.path.join(outdir, f'A_ancil_{cur_file_idx:0>3}.g3')
+                out_files.append(fname)
+                writer = core.G3Writer(fname)
+
+            m = frame_idxs == oframe_idx
+            ts = times[m]
+
+            oframe = core.G3Frame(core.G3FrameType.Scan)
+
+            i0, i1 = np.where(m)[0][[0, -1]]
+            oframe['sample_range'] = core.G3VectorInt([int(i0), int(i1+1)])
+            oframe['book_id'] = self.book_id
+
+            anc_data = core.G3TimesampleMap()
+            anc_data.times = core.G3VectorTime(ts * core.G3Units.s)
+            if az is not None:
+                anc_data['az_enc'] = core.G3VectorDouble(az[m])
+                anc_data['el_enc'] = core.G3VectorDouble(el[m])
+            if boresight is not None:
+                anc_data['boresight_enc'] = core.G3VectorDouble(boresight[m])
+            if corotation is not None:
+                anc_data['corotation_enc'] = core.G3VectorDouble(corotation[m])
+            oframe['ancil'] = anc_data
+            writer(oframe)
+            anc_frame_data.append(anc_data)
+
+            self.add_acu_summary_info(oframe, ts[0], ts[-1])
+
+        # Save this to be added to detector files
+        self.anc_frame_data = anc_frame_data
+        self.out_files = out_files
+    
+    def add_acu_summary_info(self, frame, t0, t1):
+        """
+        Adds ACU summary information to a G3Frame. This will add the following
+        info if data is present in the HK dataset::
+            - azimuth_mode: (str)
+                Azimuth_mode, pulled from the ACU summary data. `ProgramTrack`
+                means that this frame contains scan data, and `Preset` means the
+                telescope is slewing.
+            - azimuth_velocity_mean / azimuth_velocity_std: (float / float)
+                Mean and standard deviation of the az velocity (deg / sec)
+            - elevation_velocity_mean / elevaction_velocity_std: (float / float)
+                Mean and standard deviation of the el velocity (deg / sec)
+            
+        Params
+        ----------
+        frame : G3Frame
+            Frame to add data to
+        t0 : float
+            Start time of frame (unix time), inclusive
+        t1 : float
+            Stop time of frame (unix time), inclusive
+        """
+        bl = self.blocks.get('ACU_summary_output')
+        if bl.data:
+            az_mode = bl.data["Azimuth_mode"]
+            acu_summary_times = bl.times
+            m = (t0 <= acu_summary_times) & (acu_summary_times <= t1)
+            if not np.any(m):
+                frame['azimuth_mode'] = 'None'
+            elif 'ProgramTrack' in az_mode[m]:
+                frame['azimuth_mode'] = 'ProgramTrack'  # Scanning
             else:
-                # All the events are frame splits
-                frame_splits += list(event_times[spi[i]])
+                frame['azimuth_mode'] = 'Preset'  # "Slewing"
 
-        return frame_splits
+        bl = self.blocks['ACU_broadcast']
+        if bl.data:
+            az = bl.data["Corrected_Azimuth"]
+            el = bl.data["Corrected_Elevation"]
+            m = (t0 <= bl.times) & (bl.times <= t1)
+            dt = np.diff(bl.times[m]).mean()
+            az_vel = np.diff(az[m]) / dt
+            el_vel = np.diff(el[m]) / dt
 
-    def process_HK_files(self):
+            frame['azimuth_velocity_mean'] = np.mean(az_vel)
+            frame['azimuth_velocity_stdev'] = np.std(az_vel)
+            frame['elevation_velocity_mean'] = np.mean(el_vel)
+            frame['elevation_velocity_stdev'] = np.std(el_vel)
+
+        
+class SmurfStreamProcessor:
+    def __init__(self, obs_id, files, book_id, readout_ids, log=None):
+        self.files = files
+        self.obs_id = obs_id
+        self.stream_id = None
+        self.times = None
+        self.frame_idxs = None
+        self.nchans = None
+        self.nframes = None
+        self.bias_names = None
+        self.primary_names = None
+        self.timing_paradigm = None
+        self.session_id = None
+        self.slow_primary = None
+        self.sostream_version = None
+        self.readout_ids = readout_ids
+        self.out_files = []
+        self.book_id = book_id
+
+        if log is None:
+            self.log = logging.getLogger('bookbinder')
+        else:
+            self.log = log
+
+    def preprocess(self):
         """
-        Subroutine to process any provided Housekeeping (HK) files.
-
-        If no HK files provided, Bookbinder will proceed in default mode.
+        Extracts file times, nchans, and nframes from file list
         """
-        if self._hk_files is not None:
-            if not isinstance(self._hk_files, list):
-                raise TypeError("Please provide HK files in a list.")
+        if self.times is not None:  # Already preprocessed
+            return
 
-            # Chain multiple hkfile iteratables together so that
-            # by using `next` we iterate across all hk files
-            if not hasattr(self, 'hk_iter'):
-                self.hk_iter = []
-                for hkfile in self._hk_files:
-                    self.hk_iter = itertools.chain(self.hk_iter, core.G3File(hkfile))
+        self.log.info(f"Preprocessing smurf obsid {self.obs_id}")
 
-            # Initialize variables for gap detection
-            prev_frame_last_sample     = None
-            prev_frame_sample_interval = None
+        self.nframes = 0
+        ts = []
+        frame_idxs = []
+        frame_idx = 0
+        for frame in get_frame_iter(self.files):
+            if frame.type != core.G3FrameType.Scan:
+                continue
 
-            # Loop over HK frames to look for:
-            # 1. Az/El encoder data; run in default mode otherwise
-            # 2. Gaps in time between frames, which need to be flagged
-            for h in self.hk_iter:
-                if h['hkagg_type'] != 2 or ('ACU_position' not in h['block_names']):
+            # Populate attributes from the first scan frame
+            if self.nchans is None:
+                self.nchans = len(self.readout_ids)
+                self.primary_names = frame['primary'].names
+                self.bias_names = frame['tes_biases'].names
+                self.timing_paradigm = frame['timing_paradigm']
+                self.session_id = frame['session_id']
+                if 'slow_primary' in frame:
+                    self.slow_primary = frame['slow_primary']
+                self.sostream_version = frame['sostream_version']
+                self.stream_id = frame['sostream_id']
+
+            t = get_frame_times(frame)[1]
+            ts.append(t)
+            frame_idxs.append(np.full(len(t), frame_idx, dtype=np.int32))
+
+            self.nframes += 1
+            frame_idx += 1
+
+        self.times = np.hstack(ts)
+        self.frame_idxs = np.hstack(frame_idxs)
+
+    def bind(self, outdir, times, frame_idxs, file_idxs, pbar=False, ancil=None,
+             atol=1e-4):
+        """
+        Binds SMuRF data
+        
+        Params
+        ------ 
+        outdir : str
+            Output directory to put bound files
+        times : np.ndarray
+            Full array of timestamps that should be contained in the books
+        frame_idxs : np.ndarray
+            Output frame idx for each specified timestamp
+        file_idxs : np.ndarray
+            Output file idx for each specified output frame
+        pbar : bool or tqdm.tqdm
+            If True, will create a new progress bar. If False, will disable.
+            If a progress bar is passed, will use that.
+        ancil : AncilProcessor
+            Ancillary processor object (must be already bound). Ancil data will
+            be copied into the output frames.
+        atol : float
+            Absolute tolerance between smurf-timestamps and book-timestamps.
+            Samples mapped to times that are further away than atol will
+            be considered unmapped.
+        """
+        if pbar is True:
+            pbar = tqdm(total=self.nframes)
+        elif pbar is False:
+            pbar = tqdm(total=self.nframes, disable=True)
+
+        pbar.set_description(f"Binding {self.stream_id}")
+
+        self.log.info(f"Binding smurf obsid {self.obs_id}")
+
+        # Here `times` is the full array of times in the book (gapless reference
+        # times) and `self.times`` is timestamps from the pre-processed L2 data
+        # (may contain gaps)
+        sample_map = find_ref_idxs(times, self.times)
+        mapped = np.abs(times[sample_map] - self.times) < atol
+
+        oframe_idxs = frame_idxs[sample_map]  # out-frame idx for each in sample
+        oframe_idxs[~mapped] = -1 
+        _, offsets = np.unique(frame_idxs, return_index=True)
+        # Sample idx within each out-frame for every input sample
+        out_offset_idxs = sample_map - offsets[frame_idxs[sample_map]]
+
+        iframe_idxs = self.frame_idxs  #in-frame idx for each in sample
+        _, offsets = np.unique(self.frame_idxs, return_index=True)
+        # Sample idx within each in-frame for every input sample
+        in_offset_idxs = np.arange(len(self.times)) - offsets[self.frame_idxs]
+
+        # Handle file writers
+        writer = None
+        cur_file_idx = None
+        out_files = []
+
+        inframe_iter = get_frame_iter(self.files)
+        iframe, interm_frames = next_scan(inframe_iter)
+        iframe_idx = 0
+        oframe_num = 0
+        pbar.update()
+
+        for oframe_idx in np.unique(frame_idxs):
+            # Update writer
+            if file_idxs[oframe_idx] != cur_file_idx:
+                close_writer(writer)
+                cur_file_idx = file_idxs[oframe_idx]
+                fname = os.path.join(
+                    outdir, f'D_{self.stream_id}_{cur_file_idx:0>3}.g3')
+                out_files.append(fname)
+                writer = core.G3Writer(fname)
+
+            # Initialize stuff
+            m = frame_idxs == oframe_idx
+            nsamp = np.sum(m)
+            ts = times[m]
+            data = np.zeros((self.nchans, nsamp), dtype=np.int32)
+            biases = np.zeros((len(self.bias_names), nsamp), dtype=np.int32)
+            primary = np.zeros((len(self.primary_names), nsamp), dtype=np.int64)
+            filled = np.zeros(nsamp, dtype=bool)
+            
+            # Loop through in_frames filling current out_frame
+            while True:
+                # First, write any intermediate frames like observation and wiring
+                for fr in interm_frames:
+                    if 'frame_num' in fr:
+                        del fr['frame_num']
+                    fr['frame_num'] = oframe_num  # Update this so they remain ordered
+                    oframe_num += 1
+                    writer(fr)
+
+                m = (oframe_idxs == oframe_idx) & (iframe_idxs == iframe_idx)
+                outsamps = out_offset_idxs[m]
+                insamps = in_offset_idxs[m]
+
+                # Below is equivalent to:
+                #    >> data[:, outsamps] = iframe['data'][:, insamps]
+                #    >> ...
+                # However it is much faster to copy arrays by mapping contiguous
+                # chunks to contiguous chunks using slices, like:
+                #    >> arr_out[:, o0:o1] = arr_in[:, i0:i1]
+                # since numpy does not need to create a temporary copy of the
+                # data, and can just do a direct mem-map. This speeds up binding
+                # by a factory of ~4.
+                #
+                # Here we are splitting outsamps and insamps into a list
+                # of ranges where both arrays are contiguous. Then we loop
+                # through each sub-range and copy data via slicing.
+                split_idxs = 1 + np.where(
+                    (np.diff(outsamps) > 1) | (np.diff(insamps) > 1))[0]
+                outsplits = np.split(outsamps, split_idxs)
+                insplits = np.split(insamps, split_idxs)
+                for i in range(len(outsplits)):
+                    if len(insplits[i]) == 0:
+                        continue
+
+                    in0, in1 = insplits[i][0], insplits[i][-1] + 1
+                    out0, out1 = outsplits[i][0], outsplits[i][-1] + 1
+
+                    data[:, out0:out1] = iframe['data'].data[:, in0:in1]
+                    biases[:, out0:out1] = iframe['tes_biases'].data[:, in0:in1]
+                    primary[:, out0:out1] = iframe['primary'].data[:, in0:in1]
+                    filled[out0:out1] = 1
+
+                # If there are any remaining samples in the next in_frame, pull it and repeat
+                if np.any((oframe_idxs == oframe_idx) & (iframe_idxs > iframe_idx)):
+                    iframe, interm_frames = next_scan(inframe_iter)
+                    iframe_idx += 1
+                    pbar.update()
                     continue
-
-                # Check if a time gap exists since previous frame containing ACU position data
-                acu_pos_index = list(h['block_names']).index('ACU_position')
-                t = h['blocks'][acu_pos_index].times
-                if prev_frame_last_sample is not None and prev_frame_sample_interval is not None:
-                    this_frame_first_sample = t[0].time
-                    time_since_prev_frame   = this_frame_first_sample - prev_frame_last_sample
-                    if time_since_prev_frame/prev_frame_sample_interval - 1 > self.frameproc.GAP_THRESHOLD:
-                        # Add gap to internal list
-                        self.frameproc._hk_gaps.append((prev_frame_last_sample, this_frame_first_sample))
-                # update values
-                prev_frame_last_sample = t[-1].time
-                prev_frame_sample_interval = (t[-1].time - t[0].time)/(len(t)-1)
-
-                self.default_mode = False
-                self.frameproc(h)
-
-    def compile_mfile_dict(self):
-        """
-        Compile a dictionary of values to be entered into metadata file (M-file)
-
-        Returns
-        -------
-        d : dict
-            Dictionary of values
-        """
-        d = {'book_id':    self._book_id,
-             'session_id': self._session_id,
-             'start_time': self._start_time.time/core.G3Units.s,
-             'end_time':   self._end_time.time/core.G3Units.s,
-             'n_frames':   self.frame_num,
-             'sample_ranges': self.file_sample_ranges,
-             'n_samples':  self.file_sample_ranges[-1][-1]}
-
-        if self.hwp_loader is not None:
-            d.update(self.hwp_loader.get_metadata())
-
-        return d
-
-    def __call__(self):
-        """
-        Process Housekeeping (if provided) and SMuRF data.
-
-        Write out frame_splits file and metadata file before exiting.
-        """
-        # Determine if HK files contain az/el encoder data and find any gaps between HK frames
-        self.process_HK_files()
-
-        if self._hwp_root is not None:
-            self.hwp_loader = HWPLoader(self._hwp_root)
-            self.hwp_loader.load_data(self._start_time.time / core.G3Units.s, self._end_time.time / core.G3Units.s)
-            self.frameproc.hwp_loader = self.hwp_loader
-
-        # If frame splits file exists, load it; if not, determine appropriate split locations
-        if op.isfile(self._frame_splits_file):
-            self._frame_splits = [core.G3Time(t) for t in np.loadtxt(self._frame_splits_file, dtype='int', ndmin=1)]
-        else:
-            # note that find_frame_splits depends on process_HK_files having run
-            self._frame_splits = self.find_frame_splits()
-
-        for event_time in self._frame_splits:
-            self.frameproc.flush_time = event_time
-            output = []
-
-            if self.frameproc.smbundle is not None and self.frameproc.smbundle.ready(self.frameproc.flush_time):
-                output += self.frameproc.flush()
-                output = [o for o in output if len(o['signal'].times) > 0]  # Remove 0-length frames
-                if len(output) > 0:
-                    output += self.metadata
-                    self.metadata = []
-                self.write_frames(output)
-                continue
-
-            while self.frameproc.smbundle is None or not self.frameproc.smbundle.ready(self.frameproc.flush_time):
-                try:
-                    f = next(self.smurf_iter)
-                except StopIteration:
-                    # If there are remaining files, update the
-                    # SMuRF source iterator and G3 file writer
-                    if len(self._smurf_files) > 0:
-                        ifile = self._smurf_files.pop(0)
-                        if self._verbose: print(f"Bookbinding {ifile}")
-                        self.smurf_iter = core.G3File(ifile)
-                    else:
-                        # If there are no more SMuRF files, output remaining SMuRF data
-                        self.frameproc.flush_time = self.DEFAULT_TIME
-                        if self.frameproc.hkbundle is not None:
-                            # As long as pos2vel() uses np.ediff1d(), the velocity vector will be 1 sample shorter than position
-                            # Add an extra element to the end of velocity vectors to compensate
-                            self.frameproc.hkbundle.data['Azimuth_Velocity'] = np.append(self.frameproc.hkbundle.data['Azimuth_Velocity'], 0)
-                            self.frameproc.hkbundle.data['Elevation_Velocity'] = np.append(self.frameproc.hkbundle.data['Elevation_Velocity'], 0)
-                        output += self.frameproc.flush()
-                        # note that `signal` field only in the product of the FrameProcessor, not the input data
-                        output = [o for o in output if len(o['signal'].times) > 0]  # Remove 0-length frames
-                        self.write_frames(output + self.metadata)
-                        output = []
-                        self.metadata = []
-                        break
                 else:
-                    if f.type != core.G3FrameType.Scan:
-                        if self.frameproc.smbundle is None or len(self.frameproc.smbundle.times) == 0:
-                            output += [f]
-                        else:
-                            self.metadata += [f]
-                    else:
-                        output += self.frameproc(f)  # FrameProcessor returns a list of frames (can be empty)
-                        output = [o for o in output if len(o['signal'].times) > 0]  # Remove 0-length frames
-                        # Write out metadata frames only when FrameProcessor outputs one or more (scan) frames
-                        if len(output) > 0:
-                            output += self.metadata
-                            self.metadata = []
-                    self.write_frames(output)
+                    break
 
-                    # Clear output buffer after writing
-                    output = []
+            # Interpolate data where there are gaps
+            if np.all(~filled):
+                self.log.error(
+                    f"No samples properly mapped in oframe frame {oframe_idx}!"
+                    "Cannot properly interpolate for this frame."
+                )
+            elif np.any(~filled):
+                self.log.debug(
+                    f"{np.sum(~filled)} missing samples in out-frame {oframe_idx}"
+                )
+                # Missing samples at the beginning / end of a frame will be
+                # filled with the first / last sample in the frame
+                fill_value = (data[:, filled][:, 0], data[:, filled][:, -1])
+                data[:, ~filled] = interp1d(
+                    ts[filled], data[:, filled], axis=1, assume_sorted=True,
+                    kind='linear', fill_value=fill_value, bounds_error=False
+                )(ts[~filled])
 
-        self.close_file_writers()
+            m = frame_idxs == oframe_idx
+            i0, i1 = np.where(m)[0][[0, -1]]
 
-        self._frame_splits += self.frameproc._frame_splits
-        # Write frame splits file if it doesn't already exist
-        if not op.isfile(self._frame_splits_file):
-            np.savetxt(self._frame_splits_file, np.unique([int(t) for t in self._frame_splits]), fmt='%i')
+            oframe = core.G3Frame(core.G3FrameType.Scan)
 
-        # Write metadata file ('M-file')
-        with open(self._meta_file, 'w') as mfile:
-            yaml.dump(self.compile_mfile_dict(), mfile, sort_keys=False)
+            if ancil is not None:
+                t0, t1 = ts[0], ts[-1]
+                oframe['ancil'] = ancil.anc_frame_data[oframe_idx]
+                ancil.add_acu_summary_info(oframe, t0, t1)
+
+            oframe['book_id'] = self.book_id 
+            oframe['sample_range'] = core.G3VectorInt([int(i0), int(i1+1)])
+            oframe['flag_smurfgaps'] = core.G3VectorBool(~filled)
+
+            ts = core.G3VectorTime(ts * core.G3Units.s)
+            oframe['signal'] = so3g.G3SuperTimestream(self.readout_ids, ts, data)
+            oframe['primary'] = so3g.G3SuperTimestream(self.primary_names, ts, primary)
+            oframe['tes_biases'] = so3g.G3SuperTimestream(self.bias_names, ts, biases)
+            oframe['stream_id'] = self.stream_id
+            oframe['frame_num'] = oframe_num
+
+            oframe_num += 1
+            writer(oframe)
+        
+        close_writer(writer)
+        self.out_files = out_files
+
+        # In case there were remaining frames left over
+        pbar.update(self.nframes - iframe_idx - 1)
+        if pbar.n >= pbar.total:
+            pbar.close()
 
 
-##############
-# exceptions #
-##############
-
-class TimingSystemError(Exception):
-    """Exception raised when the timing system is on but timing counters are not found"""
-    pass
-
-
-#####################
-# Utility functions #
-#####################
-
-def pos2vel(p):
-    """From a given position vector, compute the velocity vector
+class BookBinder:
+    """
+    Class for combining smurf and hk L2 data to create books.
 
     Parameters
     ----------
-    p : np.ndarray
-        Position vector
+    book : sotodlib.io.imprinter.Books
+        Book object to bind
+    obsdb : dict
+        Result of imprinter.get_g3tsmurf_obs_for_book. This should be a dict
+        from obs-id to G3tSmurf Observations.
+    filedb : dict
+        Result of imprinter.get_files_for_book. This should be a dict from
+        obs-id to the list of smurf-files for that observation.
+    hkfiles : list
+        List of HK files to process.
+    max_samps_per_frame : int
+        Max number of samples per frame. This will be used to split frames
+        when the ACU data isn't present or cannot be used.
+    max_file_size : int
+        Max file size in bytes. This will be used to rotate files.
+    readout_ids : dict, optional
+        Dict of readout_ids to use for each stream_id. If provided, these
+        will be used to set the `names` in the signal frames. If not provided,
+        names will be taken from the input frames.
 
-    Returns
-    -------
-    np.ndarray
-        Velocity vector
-
+    
+    Attributes
+    -----------
+    ancil : AncilProcessor
+        Processor for ancillary data
+    streams : dict
+        Dict of SmurfStreamProcessor objects, keyed by stream_id
+    times : np.ndarray
+        Array of times for all samples in the book
+    frame_idxs : np.ndarray
+        Array of output frame indices for all samples in the book
+    file_idxs : np.ndarray
+        Array of output file indices for all output frames in the book
     """
-    return np.ediff1d(p)
+    def __init__(self, book, obsdb, filedb, data_root, readout_ids, outdir,
+                 max_samps_per_frame=50_000, max_file_size=1e9):
+        self.filedb = filedb
+        self.book = book
+        self.data_root = data_root
+        self.hk_root = os.path.join(data_root, 'hk')
+        self.meta_root = os.path.join(data_root, 'smurf')
+        self.hkfiles = get_hk_files(os.path.join(data_root, 'hk'), 
+                                    book.start.timestamp(),
+                                    book.stop.timestamp())
+        self.obsdb = obsdb
+        self.outdir = outdir
 
-def get_channel_names(s, rids=None):
-    """
-    Retrieve the channel names in a G3SuperTimestream. If provided, matches readout IDs to the
-    associated channels and returns them
+        self.max_samps_per_frame = max_samps_per_frame
+        self.max_file_size = max_file_size
 
-    Parameters
-    ----------
-    s : G3SuperTimestream
-        Timestream to query
-    rids : list
-        List of readout IDs
+        if not os.path.exists(outdir):
+            os.makedirs(outdir)
 
-    Returns
-    -------
-    list
-        List of channel names
-    """
-    if rids is not None:
-        # check compatability
-        assert len(s.names) == len(rids)
-        assert list(s.names) == sorted(s.names)  # name sure things are in our assumed order
-        # maybe more checks are needed here...
-        return rids
-    else:
-        return s.names
+        logfile = os.path.join(outdir, 'Z_bookbinder_log.txt')
+        self.log = setup_logger(logfile)
 
-def get_channel_data_from_name(s, channel_name):
-    """
-    From the channel name of a G3SuperTimestream (as listed in .names), retrieve
-    the associated data vector
+        self.ancil = AncilProcessor(self.hkfiles, book.bid, log=self.log)
+        self.streams = {}
+        for obs_id, files in filedb.items():
+            stream_id = '_'.join(obs_id.split('_')[1:-1])
+            self.streams[stream_id] = SmurfStreamProcessor(
+                obs_id, files, book.bid, readout_ids[obs_id], log=self.log
+            )
 
-    Parameters
-    ----------
-    s : G3SuperTimestream
-        Input timestream
-    channel_name : str
-        The name of the channel, exactly as listed in the .names field
+        self.times = None
+        self.frame_idxs = None
+        self.file_idxs = None
+        self.meta_files = None
+        
+    def preprocess(self):
+        """
+        Runs preprocessing steps for the book. Preprocesses smurf and ancillary
+        data. Creates full list of book-times, the output frame idx for each
+        sample, and the output file idx for each frame.
+        """
+        if self.times is not None:
+            return
 
-    Returns
-    -------
-    np.ndarray
-        Data vector associated with that channel name
-    """
-    if channel_name not in s.names:
-        raise KeyError(f"{channel_name} not found in this SuperTimestream")
-    idx = list(s.names).index(channel_name)
-    return s.data[idx]
+        for stream in self.streams.values():
+            stream.preprocess()
 
-def counters_to_timestamps(c0, c2):
-    """
-    Convert timing counter values into timestamp values (sec)
+        t0 = np.max([s.times[0] for s in self.streams.values()])
+        t1 = np.min([s.times[-1] for s in self.streams.values()])
+        ts, _ = fill_time_gaps(stream.times)
+        m = (t0 <= ts) & (ts <= t1)
+        ts = ts[m]
 
-    Copied from load_smurf.py (Jan 23, 2023)
-    """
-    s, ns = split_ts_bits(c2)
+        self.ancil.preprocess()
 
-    # Add 20 years in seconds (accounting for leap years) to handle
-    # offset between EPICS time referenced to 1990 relative to UNIX time.
-    c2 = s + ns*1e-9 + 5*(4*365 + 1)*24*60*60
-    ts = np.round(c2 - (c0 / 480000) ) + c0 / 480000
-    return ts
-
-def get_timestamps(f, use_counters):
-    """
-    Calculate the timestamp field for loaded data
-
-    Parameters
-    ----------
-    f : G3Frame
-        Input SMuRF frame containing data in G3SuperTimestream format
-    use_counters : bool
-        Whether to calcuate the timestamps from the timing counters.
-        If false, returns the times recorded in the .times field.
-
-    Returns
-    -------
-    G3VectorTime
-        The array of computed timestamps
-    """
-    if use_counters and 'primary' in f.keys():
-        counter0 = get_channel_data_from_name(f['primary'], 'Counter0')
-        if np.any(counter0):
-            counter2 = get_channel_data_from_name(f['primary'], 'Counter2')
-            timestamps = counters_to_timestamps(counter0, counter2)
-            timestamps *= core.G3Units.s
+        # Divide up frames
+        frame_splits = find_frame_splits(self.ancil)
+        if frame_splits is None:
+            frame_idxs = np.arange(len(ts)) // self.max_samps_per_frame
         else:
-            raise TimingSystemError("No timing counters found")
-    elif use_counters and 'primary' not in f.keys():
-        raise TimingSystemError("'primary' field not found")
-    else:
-        timestamps = f['data'].times
-    return core.G3VectorTime(timestamps)
+            frame_idxs = np.digitize(ts, frame_splits)
+            frame_idxs -= frame_idxs[0]
+            new_frame_idxs = frame_idxs.copy()
+            # Divide up frames that are too long
+            for fidx in np.unique(frame_idxs):
+                m = frame_idxs == fidx
+                new_frame_idxs += np.cumsum(m) // self.max_samps_per_frame
+            frame_idxs = new_frame_idxs
+
+        # Divide up files
+        samp_size = 4 # bytes
+        max_chans = np.max([s.nchans for s in self.streams.values()])
+        totsize = samp_size * max_chans * np.arange(len(ts))
+        file_idxs = []
+        for fr in np.unique(frame_idxs):
+            idx = np.where(frame_idxs == fr)[0][-1]
+            file_idxs.append(totsize[idx] // self.max_file_size)
+        file_idxs = np.array(file_idxs, dtype=int)
+
+        self.log.info("Finished preprocessing data")
+
+        self.times = ts
+        self.frame_idxs = frame_idxs
+        self.file_idxs = file_idxs
+
+    def copy_smurf_files_to_book(self):
+        """
+        Copies smurf ancillary files to an operation book.
+        """
+        if self.book.type != 'oper':
+            return
+
+        self.log.info("Copying smurf ancillary files to book")
+
+        files = []
+        for obs in self.obsdb.values():
+            files.extend(get_smurf_files(obs, self.meta_root))
+        
+        smurf_dirname = 'Z_smurf'
+        os.makedirs(os.path.join(self.outdir, smurf_dirname), exist_ok=True)
+
+        meta_files = {}
+        for f in files:
+            relpath = os.path.join(smurf_dirname, os.path.basename(f))
+            dest = os.path.join(self.outdir, relpath)
+            self.log.info(f"Copying to {dest}")
+            shutil.copyfile(f, dest)
+
+            if f.endswith('iv_analysis.npy'):
+                meta_files['iv'] = relpath
+            elif f.endswith('bg_map.npy'):
+                meta_files['bgmap'] = relpath
+            elif f.endswith('bias_step_analysis.npy'):
+                meta_files['bias_steps'] = relpath
+            elif f.endswith('take_noise.npy'):
+                meta_files['noise'] = relpath
+
+        self.meta_files = meta_files
+
+    def get_metadata(self):
+        """
+        Returns metadata dict for the book
+        """
+        self.preprocess()
+
+        meta = {}
+        meta['book_id'] = self.book.bid
+        meta['start_time'] = float(self.times[0])
+        meta['end_time'] = float(self.times[-1])
+        meta['n_frames'] = len(np.unique(self.frame_idxs))
+        meta['n_samples'] = len(self.times)
+        meta['session_id'] = self.book.bid.split('_')[1]
+
+        sample_ranges = []
+        for file_idx in np.unique(self.file_idxs):
+            fr_idxs = np.where(self.file_idxs == file_idx)[0]
+            i0 = int(np.where(self.frame_idxs == fr_idxs[0])[0][0])
+            i1 = int(np.where(self.frame_idxs == fr_idxs[-1])[0][-1])
+            sample_ranges.append([i0, i1+1])
+        meta['sample_ranges'] = sample_ranges
+
+        meta['telescope'] = self.book.tel_tube[:3].lower()
+        # parse e.g., sat1 -> st1, latc1 -> c1
+        meta['tube_slot'] = self.book.tel_tube.lower().replace("sat","satst")[3:]
+        meta['type'] = self.book.type
+        detsets = []
+        tags = []
+        for _, g3tobs in self.obsdb.items():
+            detsets.append(g3tobs.tunesets[0].name)
+            tags.append(g3tobs.tag)
+        meta['detsets'] = detsets
+
+        block = self.ancil.blocks['HWPEncoder_freq']
+        meta['hwp_freq_mean'] = None
+        meta['hwp_freq_stdev'] = None
+        t0, t1 = self.times[0], self.times[-1]
+        if 'approx_hwp_freq' in block.data:
+            hwp_freq = block.data['approx_hwp_freq']
+            m = (t0 < block.times) & (block.times < t1)
+            if m.any():
+                meta['hwp_freq_mean'] = float(np.mean(hwp_freq[m]))
+                meta['hwp_freq_stdev'] = float(np.std(hwp_freq[m]))
+        
+        meta['az_speed_mean'] = None
+        meta['az_speed_stdev'] = None
+        block = self.ancil.blocks['ACU_broadcast']
+        if 'Corrected_Azimuth' in block.data:
+            m = (t0 < block.times) & (block.times <= t1)
+            dt = np.diff(block.times[m]).mean()
+            az = block.data['Corrected_Azimuth']
+            az_speed = np.abs(np.diff(az[m]) / dt)
+            if m.any():
+                meta['az_speed_mean'] = float(np.mean(az_speed))
+                meta['az_speed_stdev'] = float(np.std(az_speed))
+
+        # make sure all tags are the same for obs in the same book
+        tags = list(set(tags))
+        assert len(tags) == 1
+        tags = tags[0].split(',')
+        # book should have at least one tag
+        assert len(tags) > 0
+        meta['subtype'] = tags[1] if len(tags) > 1 else ""
+        meta['tags'] = tags[2:]
+        meta['stream_ids'] = self.book.slots.split(',')
+        if (self.book.type == 'oper') and self.meta_files:
+            meta['meta_files'] = self.meta_files
+        return meta
+
+    def bind(self, pbar=False):
+        """
+        Binds data.
+
+        Params
+        ---------
+        pbar : bool
+            If True, will enable a progress bar.
+        """
+        self.preprocess()
+
+        self.log.info(f"Binding data to {self.outdir}")
+        if not os.path.exists(self.outdir):
+            os.makedirs(self.outdir)
+
+        # Copy smurf ancillary files if they exist
+        self.copy_smurf_files_to_book()
+
+        # Write M_file
+        mfile = os.path.join(self.outdir, 'M_index.yaml')
+        with open(mfile, 'w') as f:
+            yaml.dump(self.get_metadata(), f)
+
+        # Bind Ancil Data
+        self.ancil.bind(self.outdir, self.times, self.frame_idxs, self.file_idxs)
+        
+        tot = np.sum([s.nframes for s in self.streams.values()])
+        pbar = tqdm(total=tot, disable=(not pbar))
+        for stream in self.streams.values():
+            stream.bind(self.outdir, self.times, self.frame_idxs,
+                        self.file_idxs, pbar=pbar, ancil=self.ancil)
+
+        self.log.info("Finished binding data. Exiting.")
+        return True
+
 
 def fill_time_gaps(ts):
     """
@@ -1264,11 +801,15 @@ def fill_time_gaps(ts):
     -------------
     ts : np.ndarray
         List of timestamps of length `n`, potentially with gaps
-
+    
     Returns
     --------
     new_ts : np.ndarray
         New list of timestamps of length >= n, with gaps filled.
+    mask : np.ndarray
+        Returns a mask that tells you which elements of `new_ts` are
+        taken from real data and which are interpolated. `~mask` gives
+        the indices of the interpolated elements.
     """
     # Find indices where gaps occur and how long each gap is
     dts = np.diff(ts)
@@ -1277,8 +818,7 @@ def fill_time_gaps(ts):
     total_missing = int(np.sum(missing))
 
     # Create new array with the correct number of samples
-    # Fill with -1 assuming timestamps always > 0
-    new_ts = np.full(len(ts) + total_missing, -1, dtype=ts.dtype)
+    new_ts = np.full(len(ts) + total_missing, np.nan)
 
     # Insert old timestamps into new array with offsets that account for gaps
     offsets = np.concatenate([[0], np.cumsum(missing)])
@@ -1286,20 +826,75 @@ def fill_time_gaps(ts):
     new_ts[i0s + offsets] = ts
 
     # Use existing data to interpolate and fill holes
-    m = (new_ts < 0)
+    m = np.isnan(new_ts)
     xs = np.arange(len(new_ts))
     interp = interp1d(xs[~m], new_ts[~m])
     new_ts[m] = interp(xs[m])
 
-    return new_ts
+    return new_ts, ~m
 
-def find_missing_samples(refs, vs, atol=0.5):
+
+_primary_idx_map = {}
+def get_frame_times(frame):
     """
-    Find missing samples in a list of timestamps (vs) given a list of
-    reference timestamps (refs). The reference timestamps are assumed
-    to be the "true" timestamps, and the list of timestamps (vs) are
-    assumed to be missing some of the samples. The function returns
-    the missing samples in the list of timestamps (vs).
+    Returns timestamps for a G3Frame of detector data.
+
+    Parameters
+    --------------
+    frame : G3Frame
+        Scan frame containing detector data
+
+    Returns
+    --------------
+    high_precision : bool
+        If true, timestamps are computed from timing counters. If not, they are
+        software timestamps
+    
+    timestamps : np.ndarray
+        Array of timestamps (sec) for samples in the frame
+
+    """
+    if len(_primary_idx_map) == 0:
+        for i, name in enumerate(frame['primary'].names):
+            _primary_idx_map[name] = i
+        
+    c0 = frame['primary'].data[_primary_idx_map['Counter0']]
+    c2 = frame['primary'].data[_primary_idx_map['Counter2']]
+
+    if np.any(c0):
+        return True, counters_to_timestamps(c0, c2)
+    else:
+        return False, np.array(frame['data'].times) / core.G3Units.s
+
+
+def split_ts_bits(c):
+    """
+    Split up 64 bit to 2x32 bit
+    """
+    NUM_BITS_PER_INT = 32
+    MAXINT = (1 << NUM_BITS_PER_INT) - 1
+    a = (c >> NUM_BITS_PER_INT) & MAXINT
+    b = c & MAXINT
+    return a, b
+
+
+def counters_to_timestamps(c0, c2):
+    s, ns = split_ts_bits(c2)
+
+    # Add 20 years in seconds (accounting for leap years) to handle
+    # offset between EPOCH time referenced to 1990 relative to UNIX time.
+    c2 = s + ns*1e-9 + 5*(4*365 + 1)*24*60*60
+    ts = np.round(c2 - (c0 / 480000) ) + c0 / 480000
+    return ts
+
+
+def find_ref_idxs(refs, vs):
+    """
+    Creates a mapping from a list of timestamps (vs) to a list of reference
+    timestamps (refs). Returns an index-map of shape `vs.shape`, that maps each
+    timestamp of the array `vs` to the closest timestamp in the array `refs`.
+
+    This assumes that `refs` and `vs` are sorted in ascending order.
 
     Parameters
     ----------
@@ -1307,194 +902,229 @@ def find_missing_samples(refs, vs, atol=0.5):
         List of reference timestamps
     vs : array_like
         List of timestamps
-    atol : float
-        Absolute tolerance for the difference between the reference
 
     Returns
     -------
-    i_missing : array_like
-        indices (in refs) of the missing samples in vs
-    t_missing : array_like
-        values (in refs) of the missing samples in vs
+    idxs : array_like
+        Map of shape `vs.shape` that maps each element of `vs` to the closest
+        element in `refs`.
     """
     # Find the indices of the samples in the list of timestamps (vs)
     # that are closest to the reference timestamps
-    idx = np.searchsorted(vs, refs, side='left')
-    idx = np.clip(idx, 1, len(vs)-1)
+    idx = np.searchsorted(refs, vs, side='left')
+    idx = np.clip(idx, 1, len(refs)-1)
     # shift indices to the closest sample
-    left = vs[idx-1]
-    right = vs[idx]
-    idx -= refs - left < right - refs
-    missing = np.where(np.abs(vs[idx] - refs) > atol)[0]
-    return missing, refs[missing]
+    left = refs[idx-1]
+    right = refs[idx]
+    idx -= vs - left < right - vs
+    return idx
 
-def split_supertimestream(st, idx):
+def get_hk_files(hkdir, start, stop, tbuff=10*60):
     """
-    Split a G3SuperTimestream at a specific index
+    Gets HK files for dat between start and stop dirHWP files that may have data
+    between start and stop time
+    """
+    files = []
+    for subdir in os.listdir(hkdir):
+        try:
+            tcode = int(subdir)
+        except:
+            continue
+
+        if not start//1e5 - 1 <= tcode <= stop//1e5 + 1:
+            continue
+
+        subpath = os.path.join(hkdir, subdir)
+        files.extend([os.path.join(subpath, f) for f in os.listdir(subpath)])
+
+    files = np.array(sorted(files))
+    file_times = np.array([int(os.path.basename(f).split('.')[0]) for f in files])
+
+    m = (start-tbuff <= file_times) & (file_times < stop+tbuff)
+    if not np.any(m):
+        return []
+
+    # Add files before and after for good measure
+    fidxs = np.where(m)[0]
+    i0, i1 = fidxs[0], fidxs[-1]
+    if i0 > 0:
+        m[i0 - 1] = 1
+    if i1 < len(m) - 1:
+        m[i1 + 1] = 1
+
+    return files[m].tolist()
+
+def locate_scan_events(az, dy=0.001, min_gap=200, filter_window=100):
+    """
+    Locate places where a noisy vector t changes sign: +ve, -ve, and 0. To
+    handle noise, "zero" is defined to be a threshold region from -dy to +dy.
+    Thus t is said to have a zero-crossing if it fully crosses the threshold
+    region, with the actual crossing approximated as halfway between entering
+    and exiting the region. Entrances and exits without a crossing are also
+    identified (as "stops" and "starts" respectively).
 
     Parameters
     ----------
-    st : G3SuperTimestream
-        Input timestream to be split
-    idx : int
-        Index at which to split the timestream
+    dy : float, optional
+        Amplitude of threshold region
+    min_gap : int, optional
+        Length of gap (in samples) longer than which events are considered separate
 
     Returns
     -------
-    st0, st1 : G3SuperTimestream, G3SuperTimestream
-        Arrays formed by splitting input array st at index idx
+    events : list
+        Full list containing all zero-crossings, starts, and stops
+    starts : list
+        List of places where t exits the threshold region (subset of events)
+    stops : list
+        List of places where t enters the threshold region (subset of events)
     """
-    if idx >= len(st.times):
-        return st, None
 
-    st0 = so3g.G3SuperTimestream(st.names, st.times[:idx], st.data[:,:idx])
-    st1 = so3g.G3SuperTimestream(st.names, st.times[idx:], st.data[:,idx:])
+    offset = 0
+    if filter_window is not None:
+        boxcar = np.ones(filter_window) / filter_window
+        az = convolve(az, boxcar, mode='same')[filter_window:-filter_window]
+        offset = filter_window
+    
+    t = np.diff(az)
 
-    return st0, st1
+    tmin = np.min(t)
+    tmax = np.max(t)
 
-class HWPLoader:
+    if len(t) == 0:
+        return [], [], []
+
+    if np.sign(tmax) == np.sign(tmin):
+        return [], [], []
+
+    # If the data does not entirely cross the threshold region,
+    # do not consider it a sign change
+    if tmin > -dy and tmax < dy:
+        return [], [], []
+
+    # Find where the data crosses the lower and upper boundaries of the threshold region
+    c_lower = (np.where(np.sign(t[:-1]+dy) != np.sign(t[1:]+dy))[0] + 1)
+    c_upper = (np.where(np.sign(t[:-1]-dy) != np.sign(t[1:]-dy))[0] + 1)
+
+    # Classify each crossing as entering or exiting the threshold region
+    c_lower_enter = []; c_lower_exit = []
+    c_upper_enter = []; c_upper_exit = []
+
+    # Noise handling:
+    # If there are multiple crossings of the same boundary (upper or lower) in
+    # quick succession (i.e., less than min_gap), it is mostly likely due to
+    # noise. In this case, take the average of each group of crossings.
+    if len(c_lower) > 0:
+        spl = np.array_split(c_lower, np.where(np.ediff1d(c_lower) > min_gap)[0] + 1)
+        c_lower = np.array([int(np.ceil(np.mean(s))) for s in spl])
+        # Determine if this crossing is entering or exiting threshold region
+        for i, s in enumerate(spl):
+            if t[s[0]-1] < t[s[-1]+1]:
+                c_lower_enter.append(c_lower[i])
+            else:
+                c_lower_exit.append(c_lower[i])
+    if len(c_upper) > 0:
+        spu = np.array_split(c_upper, np.where(np.ediff1d(c_upper) > min_gap)[0] + 1)
+        c_upper = np.array([int(np.ceil(np.mean(s))) for s in spu])
+        # Determine if this crossing is entering or exiting threshold region
+        for i, s in enumerate(spu):
+            if t[s[0]-1] < t[s[-1]+1]:
+                c_upper_exit.append(c_upper[i])
+            else:
+                c_upper_enter.append(c_upper[i])
+
+    starts = np.sort(np.concatenate((c_lower_exit, c_upper_exit)))
+    stops = np.sort(np.concatenate((c_lower_enter, c_upper_enter)))
+    events = np.sort(np.concatenate((starts, stops)))
+    
+    # Look for zero-crossings
+    zc = []
+    while len(c_lower) > 0 and len(c_upper) > 0:
+        # Crossing from -ve to +ve
+        if c_lower[0] < c_upper[0]:
+            b = c_lower[c_lower < c_upper[0]]
+            zc.append(int( np.ceil(np.mean([b[-1], c_upper[0]])) ))
+            c_lower = c_lower[len(b):]
+        # Crossing from +ve to -ve
+        elif c_upper[0] < c_lower[0]:
+            b = c_upper[c_upper < c_lower[0]]
+            zc.append(int( np.ceil(np.mean([b[-1], c_lower[0]])) ))
+            c_upper = c_upper[len(b):]
+
+    # Replace all upper and lower crossings that contain a zero-crossing in
+    # between with the zero-crossing itself, but ONLY if those three events
+    # happen in quick succession (i.e., shorter than min_gap). Otherwise, they
+    # are separate events -- there is likely a stop state in between; in this
+    # case, do NOT perform the replacement.
+    for z in zc:
+        before_z = events[events < z]
+        after_z  = events[events > z]
+        if (after_z[0] - before_z[-1]) < min_gap:
+            events = np.concatenate((before_z[:-1], [z], after_z[1:]))
+
+    # Ignore first / last event if it is too close to the start or end
+    if (len(t) - events[-1] < min_gap) and events[-1] not in zc:
+        events = events[:-1]
+
+    if events[0] < min_gap:
+        events = events[1:]
+
+    return events + offset, starts + offset, stops + offset
+
+def find_frame_splits(ancil):
     """
-    Helper class to load in HWP data from data archive.
+    Determines timestamps of frame-splits from ACU data. If it cannot determine
+    frame-splits, returns None.
     """
-    def __init__(self, root_dir):
-        self.root_dir = root_dir
-        self.files = []
-        self.data = {}
-        self.times = {}
+    block = ancil.blocks['ACU_broadcast']
+    if 'Corrected_Azimuth' not in block.data:
+        return None 
 
-        self.hwp_angle_interp = None
-        self.locked_interp = None
-        self.stable_interp = None
-        self.hwp_rate_interp = None
+    az = block.data['Corrected_Azimuth']
+    idxs = locate_scan_events(az, filter_window=100)[0]
+    return block.times[idxs]
 
-    def get_files(self, start, stop):
-        """
-        Gets HWP files that may have data between start and stop time
-        """
-        files = []
-        for subdir in os.listdir(self.root_dir):
-            try:
-                tcode = int(subdir)
-            except:
-                continue
 
-            if not start//1e5 - 1 <= tcode <= stop//1e5 + 1:
-                continue
+def get_smurf_files(obs, meta_path, all_files=False):
+    """
+    Returns a list of smurf files that should be copied into a book.
 
-            subpath = os.path.join(self.root_dir, subdir)
-            files.extend([os.path.join(subpath, f) for f in os.listdir(subpath)])
+    Parameters
+    ------------
+    obs : G3tObservations
+        Observation to pull files from
+    meta_path : path
+        Smurf metadata path
+    all_files : bool
+        If true will return all found metadata files
 
-        files = np.array(sorted(files))
-        file_times = np.array([int(os.path.basename(f).split('.')[0]) for f in files])
+    Returns
+    -----------
+    files : List[path]
+        List of copyable files
+    """
 
-        m = (start <= file_times) & (file_times < stop)
-        if not np.any(m):
-            self.files = []
-            return self.files
+    def copy_to_book(file):
+        if all_files:
+            return True
+        return file.endswith('npy')
 
-        # Add files before and after for good measure
-        fidxs = np.where(m)[0]
-        i0, i1 = fidxs[0], fidxs[-1]
-        if i0 > 0:
-            m[i0 - 1] = 1
-        if i1 < len(m) - 1:
-            m[i1 + 1] = 1
+    tscode = int(obs.action_ctime//1e5)
+    files = []
 
-        self.files = files[m].tolist()
-        return self.files
+    # check adjacent folders in case action falls on a boundary
+    for tc in [tscode-1, tscode, tscode + 1]:
+        action_dir = os.path.join(
+            meta_path,
+            str(tc),
+            obs.stream_id,
+            f'{obs.action_ctime}_{obs.action_name}'
+        )
 
-    def frame_iter(self):
-        """Iterator that loops through G3 Frames of all files in self.files"""
-        return itertools.chain(*[core.G3File(f) for f in self.files])
+        if not os.path.exists(action_dir):
+            continue
 
-    def load_data(self, start, stop):
-        self.get_files(start, stop)
+        for root, _, fs in os.walk(action_dir):
+            files.extend([os.path.join(root, f) for f in fs])
 
-        data = {}  # structured data[block_name][field]
-        times = {}  # structured times[block_name]
-
-        bufftime = 30
-        for frame in self.frame_iter():
-            if frame['hkagg_type'] != so3g.HKFrameType.data:
-                continue
-
-            for bname, block in zip(frame['block_names'], frame['blocks']):
-                if bname not in data:
-                    data[bname] = {
-                        k: [] for k in block.keys()
-                    }
-                    times[bname] = []
-
-                ts = np.array(block.times) / core.G3Units.s
-
-                # Includes a buffer before and after for interpolation and timing uncertainty
-                m = ((start - bufftime) < ts) & (ts < (stop + bufftime))
-
-                for field, d in block.items():
-                    data[bname][field].append(np.array(d)[m])
-                times[bname].append(ts[m])
-
-        for bname in data:
-            for field in data[bname]:
-                data[bname][field] = np.hstack(data[bname][field])
-            times[bname] = np.hstack(times[bname])
-
-        self.data = data
-        self.times = times
-        if len(data) != 0:
-            self.hwp_angle_interp = interp1d(
-                times['fast'], np.unwrap(data['fast']['hwp_angle']),
-                bounds_error=False)
-            self.locked_interp = interp1d(
-                times['slow'], data['slow']['locked'], bounds_error=False)
-            self.stable_interp = interp1d(
-                times['slow'], data['slow']['stable'], bounds_error=False)
-            self.hwp_rate_interp = interp1d(
-                times['slow'], data['slow']['hwp_rate'], bounds_error=False)
-        else:
-            self.hwp_angle_interp = None
-            self.locked_interp = None
-            self.stable_interp = None
-            self.hwp_rate_interp = None
-
-    def get_interpolated_data(self, times):
-        if self.hwp_angle_interp is None:
-            return np.full(len(times), -HWPFlags.NO_DATA)
-
-        hwp_angle = self.hwp_angle_interp(times)
-        hwp_angle = np.mod(hwp_angle, 2*np.pi)
-
-        hwp_locked = self.locked_interp(times)
-        hwp_stable = self.stable_interp(times)
-
-        flags = np.zeros_like(hwp_angle, dtype=np.int32)
-        m = (hwp_stable == 1) & (hwp_locked != 1)
-        flags[m] = np.bitwise_or(flags[m], HWPFlags.STATIONARY)
-        m = hwp_stable != 1
-        flags[m] = np.bitwise_or(flags[m], HWPFlags.NO_DATA)
-
-        hwp_angle[flags != 0] = -flags[flags != 0]
-        return hwp_angle
-
-    def get_metadata(self):
-        meta = {}
-
-        hwp_rates = self.data['slow']['hwp_rate']
-        ts = self.times['slow']
-
-        # Report hwp rate every 10 minutes
-        fsamp = 1./np.median(np.diff(ts))
-        step = int(10 * 60 * fsamp)
-        meta['hwp_rate'] = np.array([ts, hwp_rates]).T[::step].tolist()
-
-        min_hwp_rate = 0.0001  # Hz
-        hwp_variation_thresh = 0.3  # Hz
-
-        if np.all(hwp_rates < min_hwp_rate):
-            meta['hwp_state'] = 'STATIONARY'
-        elif np.ptp(hwp_rates) > hwp_variation_thresh:
-            meta['hwp_state'] = 'CHANGING'
-        else:
-            meta['hwp_state'] = 'CONSTANT'
-
-        return meta
+    return [f for f in files if copy_to_book(f)]
